+++ frontend/src/main.js
@@ -14,7 +14,6 @@ axios.defaults.withCredentials = true
 axios.defaults.baseURL = 'http://localhost:8000/'
 
 axios.interceptors.response.use(undefined, function (error) {
-    console.log("hitting this")
     if (error) {
         const originalRequest = error.config;
         if (error.response.status === 401 && !originalRequest._retry) {