+++ frontend/src/router/index.js
@@ -8,6 +8,7 @@ import UsersView from '@/views/UsersView.vue'
 import UserView from '@/views/UserView.vue'
 import GroupsView from '@/views/GroupsView.vue'
 import GroupView from '@/views/GroupView.vue'
+import EditGroupView from '@/views/EditGroupView.vue'
 import PetsView from '@/views/PetsView.vue'
 import PetView from '@/views/PetView.vue'
 import EditPetView from '@/views/EditPetView.vue'
@@ -71,6 +72,13 @@ const router = createRouter({
       props: true
     },
     {
+      path: '/editgroup/:id',
+      name: 'EditGroup',
+      component: EditGroupView,
+      meta: { requiresAuth: true },
+      props: true
+    },
+    {
       path: '/pets',
       name: 'Pets',
       component: PetsView,
@@ -95,7 +103,6 @@ const router = createRouter({
 
 router.beforeEach((to, _, next) => {
   if (to.matched.some(record => record.meta.requiresAuth)) {
-    console.log(store.getters.isAuthenticated)
     if (store.getters.isAuthenticated) {
       next();
       return;