+++ frontend/src/views/PetView.vue
@@ -6,8 +6,30 @@
       <p><strong>DOB:</strong> {{ pet.dob }}</p>
       <p><strong>Height:</strong> {{ pet.height }}</p>
       <p><strong>Weight:</strong> {{ pet.weight }}</p>
-      <p><strong>Owners:</strong> {{ pet.users }}</p>
-      <p><strong>Groups:</strong> {{ pet.groups }}</p>
+      <p><strong>Owners:</strong> <p v-for='petUser in pet.users'>
+        <router-link :to="{name: 'User', params:{id: petUser.id}}">{{ petUser.username }}</router-link><button type="button" v-if="petUser.id == this.$store.state.users.user.id" @click="removePetUserByList(petUser.id)">X</button>
+      </p>
+      <span><button @click="editPetUsersToggle()"><span v-if="!editPetUserState">Add Owner</span><span v-else>Close</span></button></span></p>
+      <p v-if="editPetUserState"><strong>Add owner:</strong></p>
+      <p v-if="editPetUserState">
+      <form @submit="submit">
+          <div class="mb-3">
+            <multiselect
+              name="users"
+              v-model="form.users"
+              :options="petUsers"
+              :custom-label="petUser"
+              track-by="username"
+              label="username">
+            </multiselect>
+            <button type="submit" class="btn btn-primary">Submit</button>
+          </div>
+      </form>
+      </p>
+      <p v-else></p>
+      <p><strong>Groups:</strong> <p v-for='petGroup in pet.groups'>
+        <router-link :to="{name: 'Group', params:{id: petGroup.id}}">{{ petGroup.name }}</router-link>
+      </p></p>
       <p><strong>Points:</strong> {{ pet.points }}</p>
   
     <p><router-link :to="{name: 'EditPet', params:{id: pet.id}}" class="btn btn-primary">Edit</router-link></p>
@@ -17,11 +39,15 @@
   
   
   <script>
-  import { defineComponent } from 'vue';
+  import axios from 'axios';
+  import { defineComponent, ref } from 'vue';
   import { mapGetters, mapActions } from 'vuex';
+
+  import Multiselect from 'vue-multiselect';
   
   export default defineComponent({
     name: 'Pet',
+    components: { Multiselect },
     props: ['id'],
     async created() {
       try {
@@ -31,17 +57,73 @@
         this.$router.push('/');
       }
     },
+    data() {
+      return {
+        editPetUserState: false,
+        petUsers: [this.$store.state.users],
+        form: {
+          users: []
+        }
+      }
+    },
     computed: {
-      ...mapGetters({ pet: 'statePet' }),
+      ...mapGetters({ pet: 'statePet', users: 'stateUsers' }),
     },
     methods: {
-      ...mapActions(['viewPet', 'deletePet']),
+      ...mapActions(['viewPet', 'deletePet', 'getUsers', 'addPetUser', 'removePetUser']),
+      async GetAvailableUsers() {
+        let availableUserList = []
+        let petUsers = this.pet.users.map(v => v.id)
+        axios
+          .get('users')
+          .then(response => {
+            for (let i = 0; i < response.data.length; i++) {
+              let user_id = response.data[i].id
+
+              if (!new Set(petUsers).has(user_id)) {
+                availableUserList.push(response.data[i])
+              }
+            }
+            this.petUsers = availableUserList
+          })
+      },
+      petUser(option) {
+        return `${option.username}`
+      },
+      async editPetUsersToggle() {
+        this.editPetUserState = !this.editPetUserState
+        if (this.editPetUserState == true) {
+          await this.GetAvailableUsers()
+        }
+      },
+      async removePetUserByList(user_id) {
+        try {
+          let pet = {
+            id: parseInt(this.id),
+            user_id: parseInt(user_id)
+          }
+          await this.removePetUser(pet);
+          window.location.reload();
+        } catch (error) {
+          console.log(error)
+        }
+      },
       async removePet() {
         try {
-          await this.deletePet(this.id);
-          this.$router.push('/');
+          this.deletePet(this.id);
         } catch (error) {
           console.error(error);
+        }
+      },
+      async submit() {
+        try {
+          let pet = {
+            id: parseInt(this.id),
+            user_id: parseInt(this.form.users.id)
+          }
+          await this.addPetUser(pet);
+        } catch (error) {
+          console.log(error);
         }
       }
     },