+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.1] - 2023-02-04
+### Added
+- Added buttons to add and remove pets, users to and from groups
+- Added edit group page
+### Changed
+- Fixed My Pets and My Groups to accurately reflect only the pets and groups that the user owns/belongs to
+- Fixed Add Pets to Group to only include user's pets
+- Members of each item type (pet users, group pets, etc.) now link back to those items
+
 ## [0.6.0] - 2023-01-19
 ### Added
 - Added register, login pages